Add error checking for chosen config and template
Check whether they exist before trying to do things.

+++ gontributions.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/codegangsta/cli"
 	"github.com/jubalh/gontributions/gontrib"
+	"github.com/jubalh/gontributions/util"
 )
 
 const (
@@ -96,10 +97,9 @@ func run(cli *cli.Context) {
 	configPath := cli.GlobalString("config")
 	configuration, err := loadConfig(configPath)
 	if err != nil {
-		fmt.Println(err)
+		fmt.Fprintln(os.Stderr, err)
+		os.Exit(1)
 	}
-
-	contributions := gontrib.ScanContributions(configuration)
 
 	// Get users template selection
 	templateName := cli.GlobalString("template")
@@ -114,11 +114,17 @@ func run(cli *cli.Context) {
 		templatesPath = filepath.Join(templatesPath, templateFolderName)
 	}
 	absoluteTemplatePath := filepath.Join(templatesPath, templateName)
+	if !util.FileExists(absoluteTemplatePath) {
+		fmt.Fprintf(os.Stderr, "Template file %s does not exist\n", absoluteTemplatePath)
+		os.Exit(1)
+	}
+
+	contributions := gontrib.ScanContributions(configuration)
 
 	outputPath := cli.GlobalString("output")
 	f, err := os.Create(outputPath)
 	if err != nil {
-		fmt.Println(err)
+		fmt.Fprintln(os.Stderr, err)
 	}
 	defer f.Close()
 
@@ -140,7 +146,8 @@ func cmdExconf(c *cli.Context) {
 
 	text, err := json.MarshalIndent(configuration, "", "    ")
 	if err != nil {
-		fmt.Println(err)
+		fmt.Fprintln(os.Stderr, err)
 	}
+
 	fmt.Println(string(text))
 }
